requirements.txt
absl-py==0.10.0
appnope==0.1.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
boto
Brotli==1.0.9
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
cycler==0.10.0
dash==1.16.1
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.1
dash-table==4.10.1
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fastapi==0.61.1
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
google-auth==1.21.2
google-auth-oauthlib==0.4.1
grpcio==1.32.0
idna==2.10
iniconfig==1.0.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
more-itertools==8.5.0
msrest==0.6.19
nbclient==0.5.0
nbconvert==6.0.5
nbformat==5.0.7
nest-asyncio==1.4.0
notebook==6.1.4
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.4.0.44
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.10.0
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.6.1
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.0.2
python-dateutil==2.8.1
pytorch-lightning==0.9.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
sagemaker
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
Send2Trash==1.5.0
six==1.15.0
starlette==0.13.6
tensorboard==2.2.0
tensorboard-plugin-wit==1.7.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.49.0
traitlets==5.0.4
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
